+++ federation/tests/entities/activitypub/django/test_views.py
@@ -1,5 +1,5 @@
 import json
-from unittest.mock import patch, Mock
+from unittest.mock import patch
 
 import pytest
 from django.http import HttpResponse
@@ -15,14 +15,33 @@ def dummy_view(request, *args, **kwargs):
     return HttpResponse("foo")
 
 
-@method_decorator(activitypub_object_view, name='get')
-@method_decorator(activitypub_object_view, name='post')
+@activitypub_object_view
+def dummy_restricted_view(request, *args, **kwargs):
+    return HttpResponse("foo")
+
+
+@method_decorator(activitypub_object_view, name='dispatch')
 class DummyView(View):
     def get(self, request, *args, **kwargs):
         return HttpResponse("foo")
 
     def post(self, request, *args, **kwargs):
         return HttpResponse("foo")
+
+
+@method_decorator(activitypub_object_view, name='dispatch')
+class DummyRestrictedView(View):
+    def get(self, request, *args, **kwargs):
+        return HttpResponse("foo")
+
+    def post(self, request, *args, **kwargs):
+        return HttpResponse("foo")
+
+
+def dummy_get_object_function(request):
+    if request.method == 'GET':
+        return False
+    return True
 
 
 class TestActivityPubObjectView:
@@ -39,26 +58,18 @@ class TestActivityPubObjectView:
 
         assert response.content == b'foo'
 
-    @patch("federation.entities.activitypub.django.views.get_function_from_config")
-    def test_receives_messages_to_inbox(self, mock_get_config):
-        mock_func = Mock()
-        mock_get_config.return_value = mock_func
+    def test_receives_messages_to_inbox(self):
         request = RequestFactory().post("/inbox/", data='{"foo": "bar"}', content_type='application/json')
         response = dummy_view(request=request)
 
         assert response.status_code == 202
-        mock_func.assert_called_once_with(request)
 
-    @patch("federation.entities.activitypub.django.views.get_function_from_config")
-    def test_receives_messages_to_inbox__cbv(self, mock_get_config):
-        mock_func = Mock()
-        mock_get_config.return_value = mock_func
+    def test_receives_messages_to_inbox__cbv(self):
         request = RequestFactory().post("/inbox/", data='{"foo": "bar"}', content_type="application/json")
         view = DummyView.as_view()
         response = view(request=request)
 
         assert response.status_code == 202
-        mock_func.assert_called_once_with(request)
 
     @pytest.mark.parametrize('content_type', (
             'application/json', 'application/activity+json', 'application/ld+json',
@@ -88,3 +99,18 @@ class TestActivityPubObjectView:
         content = json.loads(response.content)
         assert content['name'] == 'Bob Bobértson'
         assert response['Content-Type'] == 'application/activity+json'
+
+    def test_restricted_view__denied_when_not_authorized(self):
+        request = RequestFactory().get("/", HTTP_ACCEPT='application/activity+json')
+        with patch("federation.tests.django.utils.get_object_function", new=dummy_get_object_function):
+            response = dummy_restricted_view(request=request)
+
+        assert response.status_code == 404
+
+    def test_restricted_view__denied_when_not_authorized__cbv(self):
+        request = RequestFactory().get("/", HTTP_ACCEPT='application/activity+json')
+        view = DummyRestrictedView.as_view()
+        with patch("federation.tests.django.utils.get_object_function", new=dummy_get_object_function):
+            response = view(request=request)
+
+        assert response.status_code == 404